refactor scim transform
